requirements.txt
python-ulid
numpy
nltk
scikit-learn
gensim
pandas
tensorflow
sentence_transformers
flask
openai
